+++ src/format/tests.rs
@@ -1137,6 +1137,17 @@ World"
 "#
     );
 
+    assert_format!(
+        r#"fn main() {
+  1
+  // 1
+  |> func1
+  // 2
+  |> func2
+}
+"#
+    );
+
     //
     // Let
     //